New refactor

+++ src/Gastropod.php
@@ -1,45 +1,45 @@
 <?php
 
-namespace RadFic\Gastropod\Http\Controllers;
+namespace RadFic\Gastropod;
 
-use App\Http\Controllers\Controller;
-use App\Models\User;
+use RadFic\Gastropod\GastropodAuth;
+use Illuminate\Pagination\Paginator;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Schema;
-use Illuminate\Pagination\Paginator;
 
-class BaseCrudTableController extends Controller
-{	
-    protected $class;//class name of the model	
-    protected $name;//table name
-    protected $relations;//eloquent relations of the model 
-    protected $relationsMap = [ //relations map for the rendering 
-        'relation_name' => [
-            'key' => "local_foreing_key",
-            'field' => 'related_field_to_show'
+/**
+ * Gastropod
+ *
+ * @author  Ivan Preziosi <ivan.preziosi@gmail.com>
+ *
+ */
+class Gastropod
+{
+    protected $model;//eloquent model classname
+    protected $tableName;//table name
+    protected $relations;//eloquent relations of the model
+    protected $relationsMap = [//relations map for the views
+        'profile' => [
+            'key' => "profile_id",
+            'field' => 'name',
+            'model' => Profile::class
         ]
     ];
 
-	/**
-	 * Constructor
-	 * @param $item a model instance
-	 */
-    public function __construct($item)
+    /**
+     * Constructor
+     * @param $item a model instance
+     */
+    public function __construct($modelClass, $relationsMap)
     {
-        if ($this->class == null) {
-            $this->class = Users::class;
-        }
-        $this->name = $item->getTable();
+        $this->model = $modelClass;
+        $item = new $this->model();//an empty eloquent instance
+		$this->relations = $item->getRelations();
+		$this->relationsMap = $relationsMap;
+        $this->tableName = $item->getTable();
         Paginator::useBootstrap();
     }
 
-	/**
-	 * load the eloquent relations of the model
-	 */
-    public function getRelations($item)
-    {
-        $this->relations = $item->getRelations();
-    }
 
 	/**
 	 * explore the relations map to create entries in the front end
@@ -60,25 +60,27 @@ class BaseCrudTableController extends Controller
         }
     }
 
-    public function formatShowData($showData)
+	public function formatShowData($showData)
     {
         $showData = \str_replace("[", "[<strong>", $showData);
         $showData = \str_replace("]", "</strong>]", $showData);
         return $showData;
     }
 
-
-    /**
-     * Display a listing of the resource.
+	/**
+     * Display a listing of the records in the table.
      *
+	 * @param \Illuminate\Http\Request 
      * @return \Illuminate\Http\Response
      */
-    public function index(Request $request)
+    public function index(\Illuminate\Http\Request $request)
     {
+        if (!GastropodAuth::check()) {
+            return redirect('gastropod/login');
+        }
+
         $searchkey = $request->input('search-key', session('gastropod-index-search-key'));
         $searchField = $request->input('search-field', session('gastropod-index-search-field'));
-
-
         $itemsPerPage = $request->input('ipp');
         if (!is_numeric($itemsPerPage)) {//|| $itemsPerPage<10 || $itemsPerPage>100
             $itemsPerPage = session('gastropod-index-ipp', 10);
@@ -86,7 +88,7 @@ class BaseCrudTableController extends Controller
             session(['gastropod-index-ipp' => $itemsPerPage]);
         }
 
-        $query = $this->class::query();
+        $query = $this->model::query();
 
         if ((isset($searchkey) && $searchkey != "") && (isset($searchField) && $searchField != "")) {
             $request->session()->put('gastropod-index-search-key', $searchkey);
@@ -104,20 +106,25 @@ class BaseCrudTableController extends Controller
         }
 
         $data = [
-            'name'=> $this->name,
+            'name'=> $this->tableName,
             'items' => $items
         ];
         return view('gastropod.index', $data);
     }
 
-    /**
+
+	/**
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function create()
     {
-        $item = new $this->class();
+        if (!GastropodAuth::check()) {
+            return redirect('gastropod/login');
+        }
+
+        $item = new $this->model();
         $columnNames = Schema::getColumnListing($item->getTable());
 
         $dropdowns = [];
@@ -137,58 +144,71 @@ class BaseCrudTableController extends Controller
                 }
             }
         }
-
-        //print_r(compact('dropdowns'));die();
         
         $data = [
-            'name'=> $this->name,
+            'name'=> $this->tableName,
             'columnNames' => $columnNames,
             'dropdowns' => $dropdowns
         ];
         return view('gastropod.create', $data);
     }
 
-    /**
-     * Store a newly created resource in storage.
+
+	/**
+     * Store a newly created item in the database.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(\Illuminate\Http\Request $request)
     {
-        $this->class::create($request->all());
-        return redirect()->route($this->name.'.index')
+        if (!GastropodAuth::check()) {
+            return redirect('gastropod/login');
+        }
+
+        $this->model::create($request->all());
+        return redirect()->route($this->tableName.'.index')
                         ->with('success', 'Item created successfully.');
     }
 
-    /**
-     * Display the specified resource.
+
+	/**
+     * Display the specified item.
      *
      * @param  \App\Models\User  $user
      * @return \Illuminate\Http\Response
      */
     public function show($item)
     {
-        $itemObj = $this->class::find($item)->setHidden([]);
+        if (!GastropodAuth::check()) {
+            return redirect('gastropod/login');
+        }
+
+        $itemObj = $this->model::find($item)->setHidden([]);
         foreach ($this->relationsMap as $relationName => $relationData) {
             $itemObj->$relationName;
         }
         $data = [
-            'name'=> $this->name,
+            'name'=> $this->tableName,
             'itemData' => $this->formatShowData(print_r($itemObj->toArray(), true))
         ];
         return view('gastropod.show', $data);
     }
 
-    /**
+
+	/**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\User  $user
+     * @param  Illuminate\Database\Eloquent\Model
      * @return \Illuminate\Http\Response
      */
     public function edit($item)
     {
-        $itemObj = $this->class::find($item)->setHidden([]);
+        if (!GastropodAuth::check()) {
+            return redirect('gastropod/login');
+        }
+
+        $itemObj = $this->model::find($item)->setHidden([]);
         $columnNames = Schema::getColumnListing($itemObj->getTable());
         $dropdowns = [];
         foreach ($columnNames as $columnName) {
@@ -208,40 +228,49 @@ class BaseCrudTableController extends Controller
             }
         }
         $data = [
-            'name'=> $this->name,
+            'name'=> $this->tableName,
             'item' => $itemObj->toArray(),
             'dropdowns' => $dropdowns
         ];
         return view('gastropod.edit', $data);
     }
 
-    /**
+
+	/**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\User  $user
+     * @param  \Illuminate\Http\Request
+     * @param  \Illuminate\Database\Eloquent\Model 
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $item)
+    public function update(\Illuminate\Http\Request $request, $item)
     {
-        $itemObj = $this->class::find($item)->setHidden([]);
+        if (!GastropodAuth::check()) {
+            return redirect('gastropod/login');
+        }
+
+        $itemObj = $this->model::find($item)->setHidden([]);
         $itemObj->update($request->all());
     
-        return redirect()->route($this->name.'.index')
+        return redirect()->route($this->tableName.'.index')
             ->with('success', 'Item successfully updated.');
     }
 
-    /**
+	/**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\User  $user
+     * @param  \Illuminate\Database\Eloquent\Model
      * @return \Illuminate\Http\Response
      */
     public function destroy($item)
     {
-        $itemObj = $this->class::find($item);
+        if (!GastropodAuth::check()) {
+            return redirect('gastropod/login');
+        }
+
+        $itemObj = $this->model::find($item);
         $itemObj->delete();
-        return redirect()->route($this->name.'.index')
+        return redirect()->route($this->tableName.'.index')
             ->with('success', 'Item successfully deleted.');
     }
 }
